Implement default expression annotation and include tests for it
