+++ processor/src/test/java/org/mapstruct/ap/test/source/defaultExpressions/java/DefaultExpressionDefaultValueMapper.java
@@ -18,29 +18,26 @@
  */
 package org.mapstruct.ap.test.source.defaultExpressions.java;
 
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.Mappings;
+import org.mapstruct.factory.Mappers;
+
 import java.util.Date;
+import java.util.UUID;
 
 /**
  * @author Jeffrey Smyth
  */
-public class Source {
+@Mapper ( imports = {UUID.class, Date.class } )
+public interface DefaultExpressionDefaultValueMapper {
 
-    private String id;
-    private Date date;
+    DefaultExpressionDefaultValueMapper INSTANCE = Mappers.getMapper( DefaultExpressionDefaultValueMapper.class );
 
-    public String getId() {
-        return id;
-    }
-
-    public void setId( String id ) {
-        this.id = id;
-    }
-
-    public Date getDate() {
-        return date;
-    }
-
-    public void setDate( Date date ) {
-        this.date = date;
-    }
+    @Mappings ( {
+        @Mapping ( target = "sourceId", defaultValue = "3",
+            defaultExpression = "java( UUID.randomUUID().toString() )" ),
+        @Mapping( target = "sourceDate", source = "date", defaultExpression = "java( new Date())")
+    } )
+    Target sourceToTarget( Source s );
 }
