+++ processor/src/test/java/org/mapstruct/ap/test/source/defaultExpressions/java/DefaultExpressionExpressionMapper.java
@@ -16,13 +16,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.mapstruct.ap.test.source.defaultExpressions.java.mapper;
+package org.mapstruct.ap.test.source.defaultExpressions.java;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
-import org.mapstruct.ap.test.source.defaultExpressions.java.Source;
-import org.mapstruct.ap.test.source.defaultExpressions.java.Target;
 import org.mapstruct.factory.Mappers;
 
 import java.util.Date;
@@ -31,20 +29,15 @@ import java.util.UUID;
 /**
  * @author Jeffrey Smyth
  */
-@Mapper( imports = { UUID.class, Date.class } )
-public interface SourceTargetMapper {
+@Mapper( imports = {UUID.class, Date.class } )
+public interface DefaultExpressionExpressionMapper {
 
-    SourceTargetMapper INSTANCE = Mappers.getMapper( SourceTargetMapper.class );
+    DefaultExpressionExpressionMapper INSTANCE = Mappers.getMapper( DefaultExpressionExpressionMapper.class );
 
-    @Mappings( {
-        @Mapping( target = "sourceId", source = "id", defaultExpression = "java( UUID.randomUUID().toString() )" ),
+    @Mappings ( {
+        @Mapping ( target = "sourceId", expression = "java( UUID.randomUUID().toString() )",
+            defaultExpression = "java( UUID.randomUUID().toString() )" ),
         @Mapping( target = "sourceDate", source = "date", defaultExpression = "java( new Date())")
     } )
     Target sourceToTarget( Source s );
-
-//    @Mappings( {
-//        @Mapping( target = "timeAndFormat",  expression = "java( new TimeAndFormat( s.getTime(), s.getFormat() ))"),
-//        @Mapping( target = "anotherProp", ignore = true )
-//    } )
-//    Target sourceToTargetWithMappingTarget( Source s, @MappingTarget Target t );
 }
